test(timestamp): migrate unit tests for Timestamp
